Seperate multiple functions into classes for scalability
