+++ src/batchMultiSigCall/classes/EIP712/constants.ts
