--- src/batchMultiSigCall/classes/EIP712/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from "./constants";
