--- src/batchMultiSigCall/classes/ExportFCT/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from "./EIP712StructTypes";
