+++ src/batchMultiSigCall/classes/FCTCalls/helpers.ts
