+++ src/batchMultiSigCall/classes/index.ts
@@ -1,4 +1,5 @@
 export * from "./EIP712";
+export * from "./EIP712StructTypes";
 export * from "./ExportFCT";
 export * from "./FCTCalls";
 export * from "./Options";
