+++ src/batchMultiSigCall/methods/index.ts
@@ -1,4 +1,3 @@
-export * from "./checkers";
 export * from "./FCT";
 export * from "./helpers";
 export * from "./plugins";
